feat(scoring): 段 A——军势整体乘倍率不封顶、领地恢复计分（restore-go-core-rules）
裁决 #1 / #2（2026-09-21）：

- 棋串军势 = ⌊(基础军势 + 位置加值) × 3^n / 2^n⌋，n = 倍增子数，取消封顶；
  连珠 / 协同 / 高地加值一并被倍率放大，逐棋串各取整一次。
- 总势力 = 独占空格数 + Σ棋串军势；据点分不再计入（据点代码留到段 B 删）。
- 势力 / 军势 / 倍率分子改用 BigInteger：取消封顶后 n 上界是可落子格数，
  3^81 已越过 Int128，定宽整数无法精确表示。同步改掉 determinism.md 里
  基于旧封顶前提的「大数用 Int128」与旧军势公式。
- 新增 BigIntegerJsonConverter：写出不带引号的精确十进制整数，与旧 long
  写法逐字节兼容；读入拒绝小数与指数，全程不经浮点。

测试 1245 → 1257 全绿、零警告。15 条变异全部实跑有红数记录。
既有算例按新公式逐条重算（旧值 → 新值 + 算式见 implement.md），未改期望凑绿。

另：给 trellis-implement / trellis-check 补上 mcp__codegraph__* 工具；
testing.md 新增「还原后必须刷新时间戳」——变异还原保留 mtime 会让 MSBuild
跳过重建，下一次"全绿"跑的是变异二进制。

+++ src/Siege.Core/Ai/CandidateBatch.cs
@@ -1,4 +1,5 @@
 using System.Collections.Immutable;
+using System.Numerics;
 using Siege.Core.Batch;
 using Siege.Core.Board;
 
@@ -11,7 +12,7 @@ namespace Siege.Core.Ai;
 public sealed record PointScore(Coord Coord, PieceType Type, TerrainEdit? Edit, EvaluationBreakdown Evaluation)
 {
     /// <summary>加权总分。</summary>
-    public long Total => Evaluation.Total;
+    public BigInteger Total => Evaluation.Total;
 
     /// <summary>该单点对应的暂放。</summary>
     public Placement Placement => new(Coord, Type, Edit);
@@ -23,7 +24,7 @@ public sealed record PointScore(Coord Coord, PieceType Type, TerrainEdit? Edit, 
 public sealed record CandidateBatch(ImmutableArray<Placement> Placements, EvaluationBreakdown Evaluation)
 {
     /// <summary>加权总分。</summary>
-    public long Total => Evaluation.Total;
+    public BigInteger Total => Evaluation.Total;
 
     /// <summary>落点按 <see cref="Coord"/> 字典序排好的序列（并列打破用，裁决 D4）。</summary>
     public ImmutableArray<Coord> SortedCoords => [.. Placements.Select(p => p.Coord).Order()];
